reorganize theme structure
