+++ source/common/modules/markdown-editor/theme/themes/bielefeld.ts
@@ -13,7 +13,7 @@
  */
 
 import { EditorView } from '@codemirror/view'
-import { defaultVarsDark, defaultVarsLight, type ThemeVars } from './main-override'
+import { defaultVarsDark, defaultVarsLight, type ThemeVars } from '../editor'
 
 const primaryColor = '#ffb46c'
