+++ source/common/modules/markdown-editor/theme/themes/karl-marx-stadt.ts
@@ -2,28 +2,27 @@
  * @ignore
  * BEGIN HEADER
  *
- * Contains:        Berlin Theme
+ * Contains:        Karl-Marx-Stadt Theme
  * CVM-Role:        BaseTheme
  * Maintainer:      Hendrik Erz
  * License:         GNU GPL v3
  *
- * Description:     This file contains the Berlin theme styles
+ * Description:     This file contains the Karl-Marx-Stadt theme styles
  *
  * END HEADER
  */
-
 import { EditorView } from '@codemirror/view'
-import { defaultVarsDark, defaultVarsLight, type ThemeVars } from './main-override'
+import { type ThemeVars, defaultVarsDark, defaultVarsLight } from '../editor'
 
-const primaryColor = '#1cb27e'
+const primaryColor = '#dc2d2d'
 
-const selectionLight = '#b4f0aacc'
-const selectionDark = '#5aaa50cc'
+const selectionLight = '#fbceb1'
+const selectionDark = '#a32323b3'
 
 const fontFamily = '-apple-system, BlinkMacSystemFont, "Avenir Next", Avenir, "Helvetica Neue", Helvetica, Ubuntu, Roboto, Noto, "Segoe UI", Arial, sans-serif'
 const codeFont = 'Inconsolata, monospace'
 
-const berlinVarsLight: ThemeVars = {
+export const karlMarxStadtVarsLight: ThemeVars = {
   ...defaultVarsLight,
   '--zettlr-editor-primary-color': primaryColor,
   '--zettlr-editor-secondary-color': primaryColor,
@@ -32,7 +31,7 @@ const berlinVarsLight: ThemeVars = {
   '--zettlr-editor-code-font': codeFont,
 }
 
-const berlinVarsDark: ThemeVars = {
+export const karlMarxStadtVarsDark: ThemeVars = {
   ...defaultVarsDark,
   '--zettlr-editor-primary-color': primaryColor,
   '--zettlr-editor-secondary-color': primaryColor,
@@ -41,10 +40,10 @@ const berlinVarsDark: ThemeVars = {
   '--zettlr-editor-code-font': codeFont,
 }
 
-export const themeBerlinLight = EditorView.theme({
-  '&': berlinVarsLight
+export const themeKarlMarxStadtLight = EditorView.theme({
+  '&': karlMarxStadtVarsLight
 }, { dark: false })
 
-export const themeBerlinDark = EditorView.theme({
-  '&': berlinVarsDark
+export const themeKarlMarxStadtDark = EditorView.theme({
+  '&': karlMarxStadtVarsDark
 }, { dark: true })
